// src/models/node-types/magazine-issues/types/MagazineIssueInput.ts
export type MagazineIssueInput = {
    title: string | undefined
    consecutive_number: number | undefined
    issue_number: number | undefined
    issue_year: number | undefined
    release_date: string | undefined
    single_copy_price: number | undefined
    single_copy_price_unit: string | undefined
    pages: number | undefined
}